refactor: change util class static -> component
- static사용시 멀티스레드 환경에서 병목이 발생할 수 있다.
- 싱글톤으로 관리되도록 component로 변경
