+++ src/main/java/com/example/parsing/service/ParsingService.java
@@ -2,16 +2,20 @@ package com.example.parsing.service;
 
 import com.example.parsing.domain.FormData;
 import com.example.parsing.domain.ParsingDto;
+import com.example.parsing.utils.DataParser;
+import com.example.parsing.utils.UrlConnectionJsoup;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import static com.example.parsing.utils.ParsingUtils.*;
-
 @Service
+@RequiredArgsConstructor
 public class ParsingService {
 
+    private final UrlConnectionJsoup urlConnectionJsoup;
+    private final DataParser dataParser;
+
     public ParsingDto parse(FormData formData) {
-        String data = getUrlDataJsoup(formData.getUrl());
-        data = parsing(data, formData.getType());
-        return divideUnit(data, formData.getUnit());
+        String data = urlConnectionJsoup.getUrlDataJsoup(formData.getUrl(), formData.getType());
+        return dataParser.parse(data, formData.getUnit());
     }
 }
